[#4603] Get rid of fetchLatestBlockNumberCommand
